Converted to MSVC 2005
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@1751 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/Strings.cpp
@@ -36,7 +36,7 @@ namespace CG3 {
 			u_uastrcpy(buffer, keyword);
 			keywords[entry] = new UChar[u_strlen(buffer)+1];
 			u_strcpy(keywords[entry], buffer);
-			keyword_pow[entry] = (uint32_t)pow(2, entry);
+			keyword_pow[entry] = (uint32_t)pow(2.0, (double)entry);
 			return 0;
 		}
